// src/lib/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export interface Database {
  public: {
    Tables: {
// ...
      products: {
        Row: {
          id: string
          name: string
          description: string | null
          price: number
          stock: number
          image: string | null
          images: string[] | null
          category: string
          brand: string | null
          condition: 'new' | 'used'
          specs: Json | null
          rating: number
          reviews_count: number
          created_at: string
          updated_at: string
        }
        Insert: {
          id?: string
          name: string
          description?: string | null
          price: number
          stock?: number
          image?: string | null
          images?: string[] | null
          category: string
          brand?: string | null
          condition?: 'new' | 'used'
          specs?: Json | null
          rating?: number
          reviews_count?: number
          created_at?: string
          updated_at?: string
        }
        Update: {
          id?: string
          name?: string
          description?: string | null
          price?: number
          stock?: number
          image?: string | null
          images?: string[] | null
          category?: string
          brand?: string | null
          condition?: 'new' | 'used'
          specs?: Json | null
          rating?: number
          reviews_count?: number
          created_at?: string
          updated_at?: string
        }
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
}